+++ R/tt_merge.R
@@ -57,26 +57,20 @@ tt_merge.tidytargets <- function(
     target_output <- resolved$target_output
     rm(resolved)
     
-    # Target script
-    target_script = glue("{tt_input$initialisation$store}.R")
-    
-    # Append source if any
-    write_source(user_function_source_path, target_script)
-
-    tar_append(
-        fx = tt_factory |> quote(),
-        command = wrap_quote(command),
-        target_output = target_output,
-        script = target_script,
-        ...
-    )
-
     append_step(
       tt_input,
       target_output,
-      as.list(environment())[-1] |>
-        c(list(iterate = "single"))
+      list(
+        command = command,
+        iterate = "single",
+        source = user_function_source_path,
+        factory = factory_call(
+          quote(tt_factory),
+          command = wrap_quote(command),
+          target_output = target_output,
+          ...
+        )
+      )
     )
-    
-    
+
   }